# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Restaurante CRUD",
        "description": "API para la gestión de Restaurantes: Comensales, Mesas y Reservas",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        }
    ],
    "paths": {
...
        "/api/Reserva/listReserva": {
            "get": {
                "tags": [
                    "Reserva"
                ],
                "summary": "Listar reservas",
                "description": "Retorna una lista de reservas incluyendo información del comensal y de la mesa asociada.",
                "operationId": "5c460fdd036428133ff05b0ef4eff5d3",
                "responses": {
                    "200": {
                        "description": "Lista de reservas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "rows": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "comensal_id": {
                                                        "type": "integer",
                                                        "example": 2
                                                    },
                                                    "nombre": {
                                                        "type": "string",
                                                        "example": "JOSE ZAFRA"
                                                    },
                                                    "mesa_id": {
                                                        "type": "integer",
                                                        "example": 3
                                                    },
                                                    "numero_mesa": {
                                                        "type": "string",
                                                        "example": "10"
                                                    },
                                                    "capacidad": {
                                                        "type": "integer",
                                                        "example": 4
                                                    },
                                                    "fecha": {
                                                        "type": "string",
                                                        "format": "date",
                                                        "example": "2024-04-21"
                                                    },
                                                    "hora": {
                                                        "type": "string",
                                                        "example": "19:30"
                                                    },
                                                    "numero_de_personas": {
                                                        "type": "integer",
                                                        "example": 3
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error del servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Error interno del servidor"
                                        },
                                        "data": {
                                            "properties": {
                                                "error_message": {
                                                    "type": "string",
                                                    "example": "Detalle del error"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/Reserva/crudReserva": {
            "post": {
                "tags": [
                    "Reserva"
                ],
                "summary": "Crear, editar o eliminar una reserva",
                "description": "Dependiendo del valor del parámetro 'action', se puede crear (c), actualizar (u) o eliminar (d) una reserva.\n     *     Se valida que el número de personas no supere la capacidad de la mesa.",
                "operationId": "3d5fe630c6be36ec4c322990a0da3354",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "action"
                                ],
                                "properties": {
                                    "action": {
                                        "description": "Acción a realizar: 'c' para crear, 'u' para actualizar, 'd' para eliminar",
                                        "type": "string",
                                        "example": "c"
                                    },
                                    "id": {
                                        "description": "ID de la reserva (requerido para 'u' y 'd')",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "comensal_id": {
                                        "description": "ID del comensal (requerido para 'c' y 'u')",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "mesa_id": {
                                        "description": "ID de la mesa (requerido para 'c' y 'u')",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "fecha": {
                                        "description": "Fecha de la reserva (requerido para 'c' y 'u')",
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-06-06"
                                    },
                                    "hora": {
                                        "description": "Hora de la reserva (requerido para 'c' y 'u')",
                                        "type": "string",
                                        "example": "18:00"
                                    },
                                    "numero_de_personas": {
                                        "description": "Número de personas (requerido para 'c' y 'u')",
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Operación exitosa"
                                        },
                                        "data": {
                                            "properties": {
                                                "reserva": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "comensal_id": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "mesa_id": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "fecha": {
                                                            "type": "string",
                                                            "example": "2025-06-06"
                                                        },
                                                        "hora": {
                                                            "type": "string",
                                                            "example": "18:00"
                                                        },
                                                        "numero_de_personas": {
                                                            "type": "integer",
                                                            "example": 2
                                                        },
                                                        "created_at": {
                                                            "type": "string",
                                                            "format": "date-time",
                                                            "example": "2025-04-21T14:22:00.000000Z"
                                                        },
                                                        "updated_at": {
                                                            "type": "string",
                                                            "format": "date-time",
                                                            "example": "2025-04-21T14:22:00.000000Z"
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "status": {
                                                    "type": "integer",
                                                    "example": 0
                                                },
                                                "mesa_capacidad": {
                                                    "type": "integer",
                                                    "example": 5
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error del servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Error interno del servidor"
                                        },
                                        "data": {
                                            "properties": {
                                                "error_message": {
                                                    "type": "string",
                                                    "example": "Detalle del error"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
...
}